Extracting components into separate files
DepositsInformationBlock and DepositsInformationLoading are now separate files to simplify the focus in the main deposits information block file.
